chore: remove unwired CoinbaseAccountManager/treasury path and strategy_dev lab (owner-approved)
Owner-approved deletion of two never-wired code clusters (rehabilitation
plan Wave 2, 2026-07-01, PR 9b):

Account manager cluster:
- Delete CoinbaseAccountManager (never constructed in src) + tests
- Delete `gpt-trader treasury convert|move` CLI command group; it always
  failed because no container ever provided bot.account_manager
- Delete AccountTelemetryService (never constructed in src) + tests
- Remove dead TradingBot.account_manager/account_telemetry getattr reads
  (container has no such providers); `gpt-trader account snapshot` keeps
  its exact current behavior (reports telemetry unavailable)
- Remove orphaned AccountManagerProtocol and the dead
  account_telemetry_interval knob (--account-interval was write-only)
- Remove unused BrokerBootstrapArtifacts.account_manager field

Strategy-dev lab cluster:
- Delete features/strategy_dev/lab/ and monitor/ (~2.6K unwired LOC) +
  tests; keep strategy_dev/config/ (reachable via `gpt-trader strategy`)

Decision record: docs/decisions/remove-unwired-account-manager-and-strategy-lab.md
Removals registered in docs/DEPRECATIONS.md; decision index and agent
artifacts regenerated.

No order-submission, risk, or guard logic touched.
