+++ docs/DEPRECATIONS.md
@@ -69,6 +69,10 @@ Before removing any deprecated item:
 | Yahoo data source stub (`DataSource.YAHOO`, `download_from_yahoo`) | Unreleased | Use `DataSource.COINBASE` and `download_from_coinbase`. |
 | `EventStore.events` list + `EventStore.path` JSONL alias | Unreleased | Use `EventStore.list_events()` for reads and `EventStore.root` for the storage root. |
 | TUI subsystem (`src/gpt_trader/tui/`, `gpt-trader tui` command, `--tui`/`--demo` run flags, `scripts/build_tui_css.py`, TUI CI jobs, `textual` dependency) | Unreleased | Removed; use `gpt-trader run` for the bot and `gpt-trader ideas …` for trade-idea review. See `docs/decisions/remove-tui-subsystem.md`. |
+| `gpt-trader treasury convert\|move` CLI command group (`cli/commands/treasury.py`) | Unreleased | Removed; it was never functional (no `account_manager` was ever wired, so every invocation failed). No migration path. See `docs/decisions/remove-unwired-account-manager-and-strategy-lab.md`. |
+| `CoinbaseAccountManager` (`features/brokerages/coinbase/account_manager.py`) | Unreleased | Removed as never-constructed dead code. Wire a new account-snapshot service from a fresh spec if needed. See `docs/decisions/remove-unwired-account-manager-and-strategy-lab.md`. |
+| `AccountTelemetryService` (`features/live_trade/telemetry/account.py`) + `AccountManagerProtocol` + `TradingBot.account_manager`/`account_telemetry` attributes | Unreleased | Removed as unreachable (container never provided the services). `gpt-trader account snapshot` still reports telemetry unavailable, unchanged. See `docs/decisions/remove-unwired-account-manager-and-strategy-lab.md`. |
+| `features/strategy_dev/lab/` + `features/strategy_dev/monitor/` subpackages (`Experiment*`, `ParameterGrid`, `PerformanceMonitor`, `MetricsAggregator`, `AlertManager`, …) | Unreleased | Removed as unwired dead code; `strategy_dev/config/` (used by `gpt-trader strategy`) remains. See `docs/decisions/remove-unwired-account-manager-and-strategy-lab.md`. |
 | TUI legacy preferences fallback (`config/tui_preferences.json`) | Unreleased | Use runtime preferences path or `GPT_TRADER_TUI_PREFERENCES_PATH` |
 | TUI legacy status CSS aliases (`good`/`bad`/`risk-status-*`) | Unreleased | Use `status-ok`, `status-warning`, `status-critical` |
 | TUI validation `ValidationError` alias | Unreleased | Use `FieldValidationError` |
