+++ tests/unit/gpt_trader/cli/test_cli_unit.py
@@ -24,7 +24,6 @@ def test_cli_root_help_shows_command_list(capsys):
         "coinbase",
         "controls",
         "orders",
-        "treasury",
         "report",
         "strategy",
         "optimize",
