--- tests/unit/gpt_trader/features/strategy_dev/lab/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for strategy lab module."""
