--- tests/unit/gpt_trader/features/strategy_dev/monitor/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for performance monitor module."""
