+++ var/agents/logging/event_catalog.json
@@ -20,7 +20,6 @@
       "write_to_cache"
     ],
     "monitoring": [
-      "account_telemetry",
       "api_health_guard",
       "health_assessment",
       "health_check",
@@ -44,7 +43,6 @@
       "bot_bootstrap",
       "checksum_verify",
       "collateral_update",
-      "collect_snapshot",
       "cycle",
       "daily_tracking",
       "data_quality_check",
@@ -153,23 +151,6 @@
   },
   "description": "Catalog of structured log events in GPT-Trader",
   "events": {
-    "account_telemetry": {
-      "common_fields": [
-        "stage"
-      ],
-      "components": null,
-      "levels": [
-        "DEBUG",
-        "INFO"
-      ],
-      "occurrence_count": 4,
-      "source_files": [
-        "features/live_trade/telemetry/account.py"
-      ],
-      "statuses": [
-        "disabled"
-      ]
-    },
     "add_rule": {
       "common_fields": [
         "invalid_value",
@@ -328,22 +309,6 @@
       "occurrence_count": 3,
       "source_files": [
         "features/live_trade/execution/guard_manager.py"
-      ],
-      "statuses": null
-    },
-    "collect_snapshot": {
-      "common_fields": [
-        "error_message",
-        "error_type",
-        "stage"
-      ],
-      "components": null,
-      "levels": [
-        "ERROR"
-      ],
-      "occurrence_count": 1,
-      "source_files": [
-        "features/live_trade/telemetry/account.py"
       ],
       "statuses": null
     },
@@ -2461,8 +2426,8 @@
     "details": 5,
     "environment": 12,
     "error": 60,
-    "error_message": 89,
-    "error_type": 87,
+    "error_message": 88,
+    "error_type": 86,
     "event_type": 7,
     "flatten_operation_id": 9,
     "granularity": 6,
@@ -2481,16 +2446,16 @@
     "secret_ref": 6,
     "service": 9,
     "side": 27,
-    "stage": 156,
+    "stage": 151,
     "symbol": 87
   },
   "level_distribution": {
     "CRITICAL": 2,
-    "DEBUG": 40,
-    "ERROR": 119,
-    "INFO": 91,
+    "DEBUG": 37,
+    "ERROR": 118,
+    "INFO": 90,
     "WARNING": 101
   },
-  "total_event_types": 132,
+  "total_event_types": 130,
   "version": "1.0"
 }
